requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-js-asset==3.0.1
pillow==11.0.0
sqlparse==0.5.3
